requirements.txt
requests
pandas
google-cloud-bigquery
python-dotenv
db-dtypes
pyarrow